Publish editor preview package (#19)
The goal of this change is to publish the first preview of the editor package. The CD has been updated to not only update the version of the core package (that script was converted to TS from JS during this change), but it now also sets the editor to the same version number, and adds a dependency from the editor -> core using that version number as part of the publish preparation.

The versionUp script was also updated to take into account that a package can have both a preview and latest version, and it works based off the greater of those two numbers. This is in preparation for moving from alpha to non-alpha in the future.

Also added docstrings to the public interface of the editor and updated the readme.md.

This change also tweaks the versionUp script to be consumable by other packages, although that isn't used yet.